+++ lawlibrary/static/stylesheets/lawlibrary.scss
@@ -9,26 +9,3 @@ footer {
 .btn.btn-primary {
   color: #fff;
 }
-
-
-.year-nav {
-  list-style-type: none;
-  padding-left: 0px;
-
-  li {
-    display: inline;
-    margin-right: 15px;
-
-    a,
-    a:visited {
-      color: $primary;
-      text-decoration: none;
-    }
-
-    a.active,
-    a.active:visited {
-      color: black;
-      font-weight: 500;
-    }
-  }
-}
