+++ lawlibrary/templates/lawlibrary/gazette_list.html
@@ -1,44 +1,5 @@
-{% extends "peachjam/layouts/main.html" %}
-{% load i18n %}
-{% block title %} {% trans 'Gazettes' %} {% endblock %}
+{% extends 'liiweb/gazette_list.html' %}
 
-{% block page-content %}
-  <section class="pb-5">
-    <div class="container">
-      <h1 class="my-4">{% trans 'Gazettes'%}</h1>
-
-      <div class="row">
-        {% for group in province_groups %}
-          <div class="col-lg">
-            <ul class="list-unstyled mb-0">
-              {% for locality in group %}
-                <li>
-                  <a href="{% url 'provincial_gazette_list' locality.code %}">{{ locality.name }}</a>
-                </li>
-              {% endfor %}
-            </ul>
-          </div>
-        {% endfor %}
-
-        <h4 class="sans-serif text-muted mt-3 mb-0">{{ num_gazettes }} gazette{{ num_gazettes|pluralize }}</h4>
-
-        <div class="col-md-10 mt-5 mb-5">
-          <div class="row">
-            {% for year in results %}
-              <div class="col-md-4 col-sm-6 col-xs-12 status-box">
-                <div>
-                  <h2><a class="text-decoration-none" href="{% url 'gazettes_by_year' year.year %}">{{ year.year }}</a></h2>
-                  <p class="status-up"><i class="fas fa-signal" aria-hidden="true"></i>
-                    {{ year.count }} gazette{{ year.count|pluralize }}
-                  </p>
-                </div>
-                <div data-year="{{ year.month }}" class="monthly-chart"></div>
-              </div>
-            {% endfor %}
-          </div>
-        </div>
-
-      </div>
-    </div>
-  </section>
+{% block localities %}
+  {% include 'lawlibrary/_localities.html' %}
 {% endblock %}
