+++ liiweb/static/stylesheets/liiweb.scss
@@ -1,2 +1,26 @@
 @import "stylesheets/peachjam";
 @import "components/index";
+
+
+
+.year-nav {
+  list-style-type: none;
+  padding-left: 0px;
+
+  li {
+    display: inline;
+    margin-right: 15px;
+
+    a,
+    a:visited {
+      color: $primary;
+      text-decoration: none;
+    }
+
+    a.active,
+    a.active:visited {
+      color: black;
+      font-weight: 500;
+    }
+  }
+}
