+++ liiweb/templates/liiweb/gazette_list.html
@@ -1,0 +1,36 @@
+{% extends "peachjam/layouts/main.html" %}
+{% load i18n %}
+{% block title %} {% trans 'Gazettes' %} {% endblock %}
+
+{% block page-content %}
+  <section class="pb-5">
+    <div class="container">
+      <h1 class="my-4">{% trans 'Gazettes'%}</h1>
+
+      <div class="row">
+
+      {% block localities %}
+      {% endblock %}
+
+        <h4 class="sans-serif text-muted mt-3 mb-0">{{ num_gazettes }} gazette{{ num_gazettes|pluralize }}</h4>
+
+        <div class="col-md-10 mt-5 mb-5">
+          <div class="row">
+            {% for year in results %}
+              <div class="col-md-4 col-sm-6 col-xs-12 status-box">
+                <div>
+                  <h2><a class="text-decoration-none" href="{% url 'gazettes_by_year' year.year %}">{{ year.year }}</a></h2>
+                  <p class="status-up"><i class="fas fa-signal" aria-hidden="true"></i>
+                    {{ year.count }} gazette{{ year.count|pluralize }}
+                  </p>
+                </div>
+                <div data-year="{{ year.month }}" class="monthly-chart"></div>
+              </div>
+            {% endfor %}
+          </div>
+        </div>
+
+      </div>
+    </div>
+  </section>
+{% endblock %}
