+++ liiweb/templates/liiweb/year.html
@@ -1,0 +1,51 @@
+{% extends "peachjam/layouts/main.html" %}
+{% load i18n %}
+{% block title %} {% if locality %} {{ locality }} {% endif %} {% trans 'Gazettes' %} {{ year }} {% endblock %}
+
+{% block page-content %}
+  <div class="container">
+    <header>
+      <nav aria-label="breadcrumb">
+        <ol class="breadcrumb bg-white px-0">
+          <li class="breadcrumb-item"><a href="{% url 'gazettes' %}">Gazettes</a></li>
+          {% if locality %}
+            <li class="breadcrumb-item"><a href="{% url 'provincial_gazette_list' locality.code %}">{{ locality }}</a></li>
+          {% endif %}
+          <li class="breadcrumb-item active">{{ year }}</li>
+        </ol>
+      </nav>
+
+      <h1 class="sans-serif mt-3 mb-3">{{ locality }} Gazettes {{ year }}</h1>
+
+      <div class="row">
+        <form action="/search" method="get" class="d-flex col-md-6 mb-3">
+          <input type="text" class="form-control flex-grow-1" autofocus name="q" placeholder="Search">
+          <input type="submit" class="btn btn-outline-primary ms-2" value="Search">
+        </form>
+      </div>
+
+      {% include "liiweb/_year_nav.html" %}
+
+    </header>
+
+    <div class="post-content mt-4">
+      {% for month, items in gazettes %}
+        {% if items %}
+          <h2 class="sans-serif" id="{{ forloop.counter }}">{{ month }} {{ year }}</h2>
+          <ul>
+            {% for gazette in items %}
+              <li>
+                <a href="">{{ gazette.name }}</a>
+                <span class="text-muted m-0">
+                  &nbsp;({{ gazette.size|filesizeformat }}, <a href="{{ gazette.ga_download_url }}" target="_blank">PDF</a>)
+                </span>
+              </li>
+            {% endfor %}
+          </ul>
+        {% endif %}
+      {% endfor %}
+    </div>
+
+    {% include "liiweb/_year_nav.html" %}
+
+{% endblock %}
